include tests in release source tarball
It is useful for end users to have tests to ensure the package is working correctly.

I've been working on building a package of this software to include in Debian GNU/Linux. During the package build process it is very helpful to be able to run the tests.

I also plan to run the tests frequently using the Debian continuous integration infrastructure. This will help to catch any regressions across the Python ecosystem within Debian.

Other distributions will find the tests useful for the same reason.